add quadnode test, reorganize modules
